Implement PayPal button component and integrate into order page; refactor address deletion logic and enhance error handling

+++ src/actions/address/delete-user-address.ts
@@ -2,27 +2,33 @@
 
 import prisma from "@/lib/prisma";
 
-
-
 export const deleteUserAddress = async (userId: string) => {
-
     try {
+        const address = await prisma.userAddress.findUnique({
+            where: { userId },
+        });
+
+        if (!address) {
+            return {
+                ok: false,
+                message: "La dirección no existe",
+            };
+        }
 
         await prisma.userAddress.delete({
-            where: { userId }
+            where: { userId },
         });
 
         return {
             ok: true,
-            message: "Dirección eliminada"
-        }
-        
+            message: "Dirección eliminada",
+        };
+
     } catch (error) {
         console.log(error);
         return {
             ok: false,
-            message: "Error al eliminar la dirección"
-        }        
+            message: "Error al eliminar la dirección",
+        };
     }
 };
-
